PP-3473 Separate 3ds tests
